+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+* The default text `encoder` use to encode log entries now emits the level when coloring is disabled.
 * **Deprecated** `logging.IsTraceEnabled`, define your logger and `Tracer` directly with `var zlog, tracer = logging.PackageLogger(<shortName>, "...")` instead of separately, `tracer.Enabled()` can then be used to determine if tracing should be enabled (can be enable dynamically).
 * **Deprecated** `logging.TestingOverride`, use `logging.InstantiateLoggers` directly.
 * **Deprecated** `logging.Overidde`, use `logging.InstantiateLoggers` directly and use the `logging.WithDefaultSpec` to configure the various loggers.` instead.